# docs/OpenAPI/openapi.json
{"openapi":"3.0.2","info":{"title":"dashboard_api","description":"A lightweight Cloud Optimized GeoTIFF tile server","version":"0.1.0"},"paths":{"/index.html":{"get":{"summary":"Index","description":"Demo Page.","operationId":"index_index_html_get","responses":{"200":{"description":"Successful Response","content":{"application/hmtl":{},"text/html":{"schema":{"type":"string"}}}}}}},"/":{"get":{"summary":"Index","description":"Demo Page.","operationId":"index__get","responses":{"200":{"description":"Successful Response","content":{"application/hmtl":{},"text/html":{"schema":{"type":"string"}}}}}}},"/ping":{"get":{"summary":"Ping","description":"Health Check","operationId":"ping_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/{z}/{x}/{y}@{scale}x\\.{ext}":{"get":{"tags":["tiles","tiles"],"summary":"Tile","description":"Handle /tiles requests.","operationId":"tile_v1__z___x___y___scale_x___ext__get","parameters":[{"description":"Mercator tiles's zoom level","required":true,"schema":{"title":"Z","maximum":30.0,"minimum":0.0,"type":"integer","description":"Mercator tiles's zoom level"},"name":"z","in":"path"},{"description":"Mercator tiles's column","required":true,"schema":{"title":"X","type":"integer","description":"Mercator tiles's column"},"name":"x","in":"path"},{"description":"Mercator tiles's row","required":true,"schema":{"title":"Y","type":"integer","description":"Mercator tiles's row"},"name":"y","in":"path"},{"required":true,"schema":{"title":"Scale","type":"integer"},"name":"scale","in":"path"},{"required":true,"schema":{"title":"Ext","enum":["png","npy","tif","jpg","webp"],"type":"string"},"name":"ext","in":"path"},{"description":"Cloud Optimized GeoTIFF URL.","required":true,"schema":{"title":"Url","type":"string","description":"Cloud Optimized GeoTIFF URL."},"name":"url","in":"query"},{"description":"Coma (',') delimited band indexes","required":false,"schema":{"title":"Bidx","type":"string","description":"Coma (',') delimited band indexes"},"name":"bidx","in":"query"},{"description":"Overwrite internal Nodata value.","required":false,"schema":{"title":"Nodata","anyOf":[{"type":"number"},{"type":"integer"},{"type":"string"}],"description":"Overwrite internal Nodata value."},"name":"nodata","in":"query"},{"description":"Coma (',') delimited Min,Max bounds","required":false,"schema":{"title":"Rescale","type":"string","description":"Coma (',') delimited Min,Max bounds"},"name":"rescale","in":"query"},{"required":false,"schema":{"title":"rio-color formula","type":"string"},"name":"color_formula","in":"query"},{"required":false,"schema":{"title":"rio-tiler color map name","type":"string"},"name":"color_map","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"image/png":{},"image/jpg":{},"image/webp":{},"image/tiff":{},"application/x-binary":{}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/{z}/{x}/{y}@{scale}x":{"get":{"tags":["tiles","tiles"],"summary":"Tile","description":"Handle /tiles requests.","operationId":"tile_v1__z___x___y___scale_x_get","parameters":[{"description":"Mercator tiles's zoom level","required":true,"schema":{"title":"Z","maximum":30.0,"minimum":0.0,"type":"integer","description":"Mercator tiles's zoom level"},"name":"z","in":"path"},{"description":"Mercator tiles's column","required":true,"schema":{"title":"X","type":"integer","description":"Mercator tiles's column"},"name":"x","in":"path"},{"description":"Mercator tiles's row","required":true,"schema":{"title":"Y","type":"integer","description":"Mercator tiles's row"},"name":"y","in":"path"},{"required":true,"schema":{"title":"Scale","type":"integer"},"name":"scale","in":"path"},{"description":"Output image type. Default is auto.","required":false,"schema":{"title":"Ext","enum":["png","npy","tif","jpg","webp"],"type":"string","description":"Output image type. Default is auto."},"name":"ext","in":"query"},{"description":"Cloud Optimized GeoTIFF URL.","required":true,"schema":{"title":"Url","type":"string","description":"Cloud Optimized GeoTIFF URL."},"name":"url","in":"query"},{"description":"Coma (',') delimited band indexes","required":false,"schema":{"title":"Bidx","type":"string","description":"Coma (',') delimited band indexes"},"name":"bidx","in":"query"},{"description":"Overwrite internal Nodata value.","required":false,"schema":{"title":"Nodata","anyOf":[{"type":"number"},{"type":"integer"},{"type":"string"}],"description":"Overwrite internal Nodata value."},"name":"nodata","in":"query"},{"description":"Coma (',') delimited Min,Max bounds","required":false,"schema":{"title":"Rescale","type":"string","description":"Coma (',') delimited Min,Max bounds"},"name":"rescale","in":"query"},{"required":false,"schema":{"title":"rio-color formula","type":"string"},"name":"color_formula","in":"query"},{"required":false,"schema":{"title":"rio-tiler color map name","type":"string"},"name":"color_map","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"image/png":{},"image/jpg":{},"image/webp":{},"image/tiff":{},"application/x-binary":{}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/{z}/{x}/{y}\\.{ext}":{"get":{"tags":["tiles","tiles"],"summary":"Tile","description":"Handle /tiles requests.","operationId":"tile_v1__z___x___y____ext__get","parameters":[{"description":"Mercator tiles's zoom level","required":true,"schema":{"title":"Z","maximum":30.0,"minimum":0.0,"type":"integer","description":"Mercator tiles's zoom level"},"name":"z","in":"path"},{"description":"Mercator tiles's column","required":true,"schema":{"title":"X","type":"integer","description":"Mercator tiles's column"},"name":"x","in":"path"},{"description":"Mercator tiles's row","required":true,"schema":{"title":"Y","type":"integer","description":"Mercator tiles's row"},"name":"y","in":"path"},{"required":true,"schema":{"title":"Ext","enum":["png","npy","tif","jpg","webp"],"type":"string"},"name":"ext","in":"path"},{"description":"Tile size scale. 1=256x256, 2=512x512...","required":false,"schema":{"title":"Scale","exclusiveMaximum":4.0,"exclusiveMinimum":0.0,"type":"integer","description":"Tile size scale. 1=256x256, 2=512x512...","default":1},"name":"scale","in":"query"},{"description":"Cloud Optimized GeoTIFF URL.","required":true,"schema":{"title":"Url","type":"string","description":"Cloud Optimized GeoTIFF URL."},"name":"url","in":"query"},{"description":"Coma (',') delimited band indexes","required":false,"schema":{"title":"Bidx","type":"string","description":"Coma (',') delimited band indexes"},"name":"bidx","in":"query"},{"description":"Overwrite internal Nodata value.","required":false,"schema":{"title":"Nodata","anyOf":[{"type":"number"},{"type":"integer"},{"type":"string"}],"description":"Overwrite internal Nodata value."},"name":"nodata","in":"query"},{"description":"Coma (',') delimited Min,Max bounds","required":false,"schema":{"title":"Rescale","type":"string","description":"Coma (',') delimited Min,Max bounds"},"name":"rescale","in":"query"},{"required":false,"schema":{"title":"rio-color formula","type":"string"},"name":"color_formula","in":"query"},{"required":false,"schema":{"title":"rio-tiler color map name","type":"string"},"name":"color_map","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"image/png":{},"image/jpg":{},"image/webp":{},"image/tiff":{},"application/x-binary":{}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/{z}/{x}/{y}":{"get":{"tags":["tiles","tiles"],"summary":"Tile","description":"Handle /tiles requests.","operationId":"tile_v1__z___x___y__get","parameters":[{"description":"Mercator tiles's zoom level","required":true,"schema":{"title":"Z","maximum":30.0,"minimum":0.0,"type":"integer","description":"Mercator tiles's zoom level"},"name":"z","in":"path"},{"description":"Mercator tiles's column","required":true,"schema":{"title":"X","type":"integer","description":"Mercator tiles's column"},"name":"x","in":"path"},{"description":"Mercator tiles's row","required":true,"schema":{"title":"Y","type":"integer","description":"Mercator tiles's row"},"name":"y","in":"path"},{"description":"Tile size scale. 1=256x256, 2=512x512...","required":false,"schema":{"title":"Scale","exclusiveMaximum":4.0,"exclusiveMinimum":0.0,"type":"integer","description":"Tile size scale. 1=256x256, 2=512x512...","default":1},"name":"scale","in":"query"},{"description":"Output image type. Default is auto.","required":false,"schema":{"title":"Ext","enum":["png","npy","tif","jpg","webp"],"type":"string","description":"Output image type. Default is auto."},"name":"ext","in":"query"},{"description":"Cloud Optimized GeoTIFF URL.","required":true,"schema":{"title":"Url","type":"string","description":"Cloud Optimized GeoTIFF URL."},"name":"url","in":"query"},{"description":"Coma (',') delimited band indexes","required":false,"schema":{"title":"Bidx","type":"string","description":"Coma (',') delimited band indexes"},"name":"bidx","in":"query"},{"description":"Overwrite internal Nodata value.","required":false,"schema":{"title":"Nodata","anyOf":[{"type":"number"},{"type":"integer"},{"type":"string"}],"description":"Overwrite internal Nodata value."},"name":"nodata","in":"query"},{"description":"Coma (',') delimited Min,Max bounds","required":false,"schema":{"title":"Rescale","type":"string","description":"Coma (',') delimited Min,Max bounds"},"name":"rescale","in":"query"},{"required":false,"schema":{"title":"rio-color formula","type":"string"},"name":"color_formula","in":"query"},{"required":false,"schema":{"title":"rio-tiler color map name","type":"string"},"name":"color_map","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"image/png":{},"image/jpg":{},"image/webp":{},"image/tiff":{},"application/x-binary":{}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/tilejson.json":{"get":{"tags":["metadata"],"summary":"Tilejson","description":"Handle /tilejson.json requests.","operationId":"tilejson_v1_tilejson_json_get","parameters":[{"description":"Cloud Optimized GeoTIFF URL.","required":true,"schema":{"title":"Url","type":"string","description":"Cloud Optimized GeoTIFF URL."},"name":"url","in":"query"},{"description":"Output image type. Default is auto.","required":false,"schema":{"title":"Tile Format","enum":["png","npy","tif","jpg","webp"],"type":"string","description":"Output image type. Default is auto."},"name":"tile_format","in":"query"},{"description":"Tile size scale. 1=256x256, 2=512x512...","required":false,"schema":{"title":"Tile Scale","exclusiveMaximum":4.0,"exclusiveMinimum":0.0,"type":"integer","description":"Tile size scale. 1=256x256, 2=512x512...","default":1},"name":"tile_scale","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TileJSON"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/bounds":{"get":{"tags":["metadata"],"summary":"Bounds","description":"Handle /bounds requests.","operationId":"bounds_v1_bounds_get","parameters":[{"description":"Cloud Optimized GeoTIFF URL.","required":true,"schema":{"title":"Url","type":"string","description":"Cloud Optimized GeoTIFF URL."},"name":"url","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/metadata":{"get":{"tags":["metadata"],"summary":"Metadata","description":"Handle /metadata requests.","operationId":"metadata_v1_metadata_get","parameters":[{"description":"Cloud Optimized GeoTIFF URL.","required":true,"schema":{"title":"Url","type":"string","description":"Cloud Optimized GeoTIFF URL."},"name":"url","in":"query"},{"description":"Coma (',') delimited band indexes","required":false,"schema":{"title":"Bidx","type":"string","description":"Coma (',') delimited band indexes"},"name":"bidx","in":"query"},{"description":"Overwrite internal Nodata value.","required":false,"schema":{"title":"Nodata","anyOf":[{"type":"number"},{"type":"integer"},{"type":"string"}],"description":"Overwrite internal Nodata value."},"name":"nodata","in":"query"},{"required":false,"schema":{"title":"Pmin","type":"number","default":2.0},"name":"pmin","in":"query"},{"required":false,"schema":{"title":"Pmax","type":"number","default":98.0},"name":"pmax","in":"query"},{"required":false,"schema":{"title":"Max Size","type":"integer","default":1024},"name":"max_size","in":"query"},{"required":false,"schema":{"title":"Histogram Bins","type":"integer","default":20},"name":"histogram_bins","in":"query"},{"description":"Coma (',') delimited Min,Max bounds","required":false,"schema":{"title":"Histogram Range","type":"string","description":"Coma (',') delimited Min,Max bounds"},"name":"histogram_range","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/WMTSCapabilities.xml":{"get":{"tags":["OGC"],"summary":"Wtms","description":"Wmts endpoit.","operationId":"wtms_v1_WMTSCapabilities_xml_get","parameters":[{"description":"Cloud Optimized GeoTIFF URL.","required":true,"schema":{"title":"Url","type":"string","description":"Cloud Optimized GeoTIFF URL."},"name":"url","in":"query"},{"description":"Output image type. Default is png.","required":false,"schema":{"title":"Tile Format","enum":["png","npy","tif","jpg","webp"],"type":"string","description":"Output image type. Default is png.","default":"png"},"name":"tile_format","in":"query"},{"description":"Tile size scale. 1=256x256, 2=512x512...","required":false,"schema":{"title":"Tile Scale","exclusiveMaximum":4.0,"exclusiveMinimum":0.0,"type":"integer","description":"Tile size scale. 1=256x256, 2=512x512...","default":1},"name":"tile_scale","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/xml":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"TileJSON":{"title":"TileJSON","required":["tiles","center"],"type":"object","properties":{"tilejson":{"title":"Tilejson","type":"string","default":"2.2.0"},"name":{"title":"Name","type":"string"},"description":{"title":"Description","type":"string"},"version":{"title":"Version","type":"string","default":"1.0.0"},"attribution":{"title":"Attribution","type":"string"},"template":{"title":"Template","type":"string"},"legend":{"title":"Legend","type":"string"},"scheme":{"title":"Scheme","type":"string","default":"xyz"},"tiles":{"title":"Tiles","type":"array","items":{"type":"string"}},"grids":{"title":"Grids","type":"array","items":{"type":"string"}},"data":{"title":"Data","type":"array","items":{"type":"string"}},"minzoom":{"title":"Minzoom","maximum":30.0,"minimum":0.0,"type":"integer","default":0},"maxzoom":{"title":"Maxzoom","maximum":30.0,"minimum":0.0,"type":"integer","default":30},"bounds":{"title":"Bounds","type":"array","items":{"type":"number"},"default":[-180,-90,180,90]},"center":{"title":"Center","type":"array","items":[{"type":"number"},{"type":"number"},{"type":"integer"}]}},"description":"TileJSON model.\n\nBased on https://github.com/mapbox/tilejson-spec/tree/master/2.2.0"},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}